requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openpyxl
pdfplumber
python-multipart
httpx
anthropic
tavily-python
python-dotenv
apscheduler
aiofiles
